+++ CHANGELOG.md
@@ -26,6 +26,7 @@ To release a new version (e.g. from `1.0.0` -> `2.0.0`):
 * Fix `response_curves` when optimized data timeframe is outside of modeled
   data timeframe.
 * Add `lognormal_dist_from_mean_std` helper function.
+* Add `use_kpi` arg to `output_model_results_summary`.
 
 ## [1.2.0] - 2025-09-04